Refactoring logger into common class to enable reuse.
