+++ tests/OpenCloud/Tests/MockLogger.php
@@ -1,0 +1,35 @@
+<?php
+/**
+ * Copyright 2012-2014 Rackspace US, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace OpenCloud\Tests;
+
+use OpenCloud\Common\Log\Logger;
+
+class MockLogger extends Logger
+{
+    protected $criticalLogMessage;
+
+    public function critical($message, array $context = array())
+    {
+        ++$this->criticalLogMessage;
+    }
+
+    public function getCriticalLogMessage()
+    {
+        return $this->criticalLogMessage;
+    }
+}
